moved data import template to separate file
