+++ server/modules/dataImport/service/DataImportTemplateService.js
@@ -1,0 +1,58 @@
+import { PointFactory, Points } from '@openforis/arena-core'
+
+import { CsvExportModel } from '@common/model/csvExport'
+
+import * as Survey from '@core/survey/survey'
+import * as NodeDef from '@core/survey/nodeDef'
+import * as DateUtils from '@core/dateUtils'
+
+import * as CSVWriter from '@server/utils/file/csvWriter'
+
+import * as SurveyManager from '@server/modules/survey/manager/surveyManager'
+
+const valuesByNodeDefType = {
+  [NodeDef.nodeDefType.boolean]: () => true,
+  [NodeDef.nodeDefType.code]: () => 'CATEGORY_CODE',
+  [NodeDef.nodeDefType.coordinate]: () =>
+    Points.toString(PointFactory.createInstance({ srs: 'EPSG:4326', x: 41.8830209, y: 12.4879562 })),
+  [NodeDef.nodeDefType.date]: () => DateUtils.formatDateISO(new Date()),
+  [NodeDef.nodeDefType.decimal]: () => 123.45,
+  [NodeDef.nodeDefType.file]: () => '', // TODO support file attribute import
+  [NodeDef.nodeDefType.integer]: () => 123,
+  [NodeDef.nodeDefType.taxon]: () => 'TAXON_CODE',
+  [NodeDef.nodeDefType.text]: () => 'Text',
+  [NodeDef.nodeDefType.time]: () => {
+    const now = new Date()
+    return DateUtils.formatTime(now.getHours(), now.getMinutes())
+  },
+}
+
+const extractDataImportTemplate = async ({ surveyId, cycle, entityDefUuid }) => {
+  const survey = await SurveyManager.fetchSurveyAndNodeDefsBySurveyId({ surveyId, cycle })
+  const entityDef = Survey.getNodeDefByUuid(entityDefUuid)(survey)
+  const exportModel = new CsvExportModel({
+    survey,
+    nodeDefContext: entityDef,
+    options: { includeFiles: false, includeCategoryItemsLabels: false, includeTaxonScientificName: false },
+  })
+  const dataRow = exportModel.columns.reduce((acc, column) => {
+    const { header, nodeDef, valueProp } = column
+    let value
+    if (nodeDef) {
+      value = valuesByNodeDefType[NodeDef.getType(nodeDef)]({ valueProp })
+    } else {
+      value = ''
+    }
+    return { ...acc, [header]: value }
+  }, {})
+  return dataRow
+}
+
+const writeDataImportTemplateToStream = async ({ surveyId, cycle, entityDefUuid, outputStream }) => {
+  const dataImportTemplate = await extractDataImportTemplate({ surveyId, cycle, entityDefUuid })
+  await CSVWriter.writeToStream(outputStream, [dataImportTemplate])
+}
+
+export const DataImportTemplateService = {
+  writeDataImportTemplateToStream,
+}
